// src/lib/api/model/onboardingRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Unplan API
 * Unplan Server API Documentation
 * OpenAPI spec version: v1.0.0
 */
import type { UpdateMethods } from './updateMethods';
import type { UpdateConditions } from './updateConditions';
import type { BiorhythmRequest } from './biorhythmRequest';
import type { OnboardingRequestTransportationsItem } from './onboardingRequestTransportationsItem';

export interface OnboardingRequest {
  recovery: UpdateMethods;
  sleep_condition: UpdateConditions;
  biorhythm: BiorhythmRequest;
  transportations?: OnboardingRequestTransportationsItem[];
}